Alpha of MT split into isothermal and thermal
